requirements.txt
tensorflow
pandas
numpy
scipy
matplotlib
seaborn
dvc
notebook
python-box==6.0.2
PyYAML
tqdm
ensure==1.0.2
joblib
Flask
Flask-Cors
-e .
